requirements.txt
logzero
pyqt5

python-dotenv
